handles floating numbers in expression.

+++ src/digvijag/evaluator/cli/ExpressionEvaluator.java
@@ -5,7 +5,7 @@ import digvijag.evaluator.lib.ExpressionEvaluatorLibrary;
 public class ExpressionEvaluator {
     public static void main(String[] args) {
         String expression = args[0];
-        int result = new ExpressionEvaluatorLibrary().evaluate(expression);
+        double result = new ExpressionEvaluatorLibrary().evaluate(expression);
         System.out.println(result);
     }
 }
