+++ src/digvijag/evaluator/lib/SingleOperationEvaluator.java
@@ -4,7 +4,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 interface Evaluable {
-    int evaluate(int leftOperand, int rightOperand);
+    double evaluate(double leftOperand, double rightOperand);
 }
 
 public class SingleOperationEvaluator {
@@ -15,59 +15,59 @@ public class SingleOperationEvaluator {
 
         operationMap.put("+", new Evaluable() {
             @Override
-            public int evaluate(int leftOperand, int rightOperand) {
+            public double evaluate(double leftOperand, double rightOperand) {
                 return leftOperand + rightOperand;
             }
         });
 
         operationMap.put("-", new Evaluable() {
             @Override
-            public int evaluate(int leftOperand, int rightOperand) {
+            public double evaluate(double leftOperand, double rightOperand) {
                 return leftOperand - rightOperand;
             }
         });
 
         operationMap.put("*", new Evaluable() {
             @Override
-            public int evaluate(int leftOperand, int rightOperand) {
+            public double evaluate(double leftOperand, double rightOperand) {
                 return leftOperand * rightOperand;
             }
         });
 
         operationMap.put("/", new Evaluable() {
             @Override
-            public int evaluate(int leftOperand, int rightOperand) {
+            public double evaluate(double leftOperand, double rightOperand) {
                 return leftOperand / rightOperand;
             }
         });
 
         operationMap.put("^", new Evaluable() {
             @Override
-            public int evaluate(int leftOperand, int rightOperand) {
-                return (int) Math.pow(leftOperand, rightOperand);
+            public double evaluate(double leftOperand, double rightOperand) {
+                return Math.pow(leftOperand, rightOperand);
             }
         });
 
         this.operationMap = operationMap;
     }
 
-    public int evaluateOperation(String expression) {
+    public double evaluateOperation(String expression) {
         String operations = "+-*/^";
         String operation = null;
-        int leftOperand = 0, rightOperand = 0;
+        double leftOperand = 0, rightOperand = 0;
         String tempOperand;
         for (int i = 0; i < expression.length(); i++) {
             String currentCharacter = Character.toString(expression.charAt(i));
             if (operations.contains(currentCharacter)) {
                 tempOperand = expression.substring(0, i).trim();
-                leftOperand = Integer.parseInt(tempOperand);
+                leftOperand = Double.parseDouble(tempOperand);
                 tempOperand = expression.substring(i + 1, expression.length()).trim();
-                rightOperand = Integer.parseInt(tempOperand);
+                rightOperand = Double.parseDouble(tempOperand);
                 operation = currentCharacter;
             }
         }
         expression = expression.trim();
-        if (operation == null) return Integer.parseInt(expression);
+        if (operation == null) return Double.parseDouble(expression);
 
         return operationMap.get(operation).evaluate(leftOperand, rightOperand);
     }
