+++ test/digvijag/lib/BracketsTest.java
@@ -10,7 +10,7 @@ public class BracketsTest {
     public void test_solve_expression_with_single_bracket() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 5;
-        int actual = evaluator.evaluate("(3+2)");
+        int actual = (int) evaluator.evaluate("(3+2)");
         assertEquals(expected, actual);
     }
 
@@ -18,7 +18,7 @@ public class BracketsTest {
     public void test_solve_expression_with_two_brackets() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 10;
-        int actual = evaluator.evaluate("(3+2)+(3+2)");
+        int actual = (int) evaluator.evaluate("(3+2)+(3+2)");
         assertEquals(expected, actual);
     }
 
@@ -26,7 +26,7 @@ public class BracketsTest {
     public void test_solve_expression_solves_brackets_first() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 13;
-        int actual = evaluator.evaluate("3+(5*2)");
+        int actual = (int) evaluator.evaluate("3+(5*2)");
         assertEquals(expected, actual);
     }
 
@@ -34,7 +34,7 @@ public class BracketsTest {
     public void test_solve_expression_solves_nested_brackets() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 10;
-        int actual = evaluator.evaluate("(3+(5+2))");
+        int actual = (int) evaluator.evaluate("(3+(5+2))");
         assertEquals(expected, actual);
     }
 
@@ -42,7 +42,7 @@ public class BracketsTest {
     public void test_solve_expression_solves_multiple_nested_brackets() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 7;
-        int actual = evaluator.evaluate("(3+(2+1)+(1))");
+        int actual = (int) evaluator.evaluate("(3+(2+1)+(1))");
         assertEquals(expected, actual);
     }
 }
