+++ test/digvijag/lib/ExpressionEvaluatorLibraryTest.java
@@ -10,7 +10,7 @@ public class ExpressionEvaluatorLibraryTest {
     public void test_addition_of_two_integers() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 5;
-        int actual = evaluator.evaluate("2+3");
+        int actual = (int) evaluator.evaluate("2+3");
         assertEquals(expected, actual);
     }
 
@@ -18,7 +18,7 @@ public class ExpressionEvaluatorLibraryTest {
     public void test_subtraction_of_two_integers() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 1;
-        int actual = evaluator.evaluate("3-2");
+        int actual = (int) evaluator.evaluate("3-2");
         assertEquals(expected, actual);
     }
 
@@ -26,7 +26,7 @@ public class ExpressionEvaluatorLibraryTest {
     public void test_multiplication_of_two_integers() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 6;
-        int actual = evaluator.evaluate("3*2");
+        int actual = (int) evaluator.evaluate("3*2");
         assertEquals(expected, actual);
     }
 
@@ -34,7 +34,7 @@ public class ExpressionEvaluatorLibraryTest {
     public void test_division_of_two_integers() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 3;
-        int actual = evaluator.evaluate("6/2");
+        int actual = (int) evaluator.evaluate("6/2");
         assertEquals(expected, actual);
     }
 
@@ -42,7 +42,7 @@ public class ExpressionEvaluatorLibraryTest {
     public void test_power_of_two_integers() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 4;
-        int actual = evaluator.evaluate("2^2");
+        int actual = (int) evaluator.evaluate("2^2");
         assertEquals(expected, actual);
     }
 
@@ -50,7 +50,7 @@ public class ExpressionEvaluatorLibraryTest {
     public void test_when_no_operation_is_specified() {
         ExpressionEvaluatorLibrary evaluator = new ExpressionEvaluatorLibrary();
         int expected = 2;
-        int actual = evaluator.evaluate("2");
+        int actual = (int) evaluator.evaluate("2");
         assertEquals(expected, actual);
     }
